segregate functions from the components
